+++ src/main/scala/com/atomist/source/git/github/domain/Repository.scala
@@ -1,0 +1,13 @@
+package com.atomist.source.git.github.domain
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+case class Repository(id: Long,
+                      name: String,
+                      description: String,
+                      url: String,
+                      @JsonProperty("forks_count") forksCount: Int,
+                      owner: RepoOwner) {
+
+  def ownerName: String = owner.login
+}
